requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
numpy==1.26.4
tensorflow==2.10.0
werkzeug==3.0.1